+++ src/main/java/com/kodilla/backend/order/decorator/OrderInterface.java
@@ -2,7 +2,7 @@ package com.kodilla.backend.order.decorator;
 
 public interface OrderInterface {
     Double getCost();
-    Double getDistance();
+    Integer getDistance();
     String getDescription();
 
 }
